Split up profiler tests, simplify exports
