+++ test/unit/contractProfiler.test.js
@@ -1,17 +1,13 @@
 const assert = require('assert')
-const deepEqual = require('deep-equal')
 const fs = require('fs')
 
-const { contractProfiler } = require('../../lib/profilers/contractProfiler')
-const { systemProfiler } = require('../../lib/profilers/systemProfiler')
-const { derivedProfiler } = require('../../lib/profilers/derivedProfiler')
+const contractProfiler = require('../../lib/profilers/contractProfiler')
+const systemProfiler = require('../../lib/profilers/systemProfiler')
 
 const featureful = './test/contracts/Featureful.sol'
 
 
-// Manually define the expected output for profiles
-
-// state variable profiles to match
+// Manually define state variable profiles to match
 let valueTypeProfile = { name: 'a', visibility: 'default', typeInfo: { type: 'bytes32' } }
 let userDefinedTypeProfile = { name: 'aStruct', visibility: 'default', typeInfo: { type: 'MyType'} }
 let fixedArrayProfile = { name: 'fixedArray', visibility: 'default', typeInfo: { type: 'array', fixed: true, length: { type: 'NumberLiteral', number: '21', subdenomination: null},baseType: 'bytes32'}}
@@ -71,59 +67,6 @@ describe('Profilers', function() {
       assert.deepEqual(profiles[1].functionProfiles[1] , fooProfile)
       assert.deepEqual(profiles[1].functionProfiles[2] , barProfile)
       assert.deepEqual(profiles[1].functionProfiles[3] , fallbackProfile)
-    })
-  })
-
-  describe('systemProfiler', function() {
-    let contractProfiles, inheritanceGraph
-    // we'll use openzeppelin's ERC20 dir as a reasonable complex system for testing
-    const tokensDir = './node_modules/openzeppelin-solidity/contracts/token/ERC20/'
-    // an alphabetically sorted list of all declared contract names found by grepping through the folder:
-    let declaredTokens = [ 
-      'PausableToken',
-      'StandardToken',
-      'MintableToken',
-      'BasicToken',
-      'DetailedERC20',
-      'ERC20',
-      'ERC20Basic',
-      'BurnableToken',
-      'StandardBurnableToken',
-      'TokenVesting',
-      'RBACMintableToken',
-      'CappedToken',
-      // 'uncomment to make me fail'
-    ]
-
-    before(function() {
-      let tokensFiles = fs.readdirSync(tokensDir).map(entry => `${tokensDir}${entry}`)
-      let systemProfile = systemProfiler(tokensFiles)
-      contractProfiles = systemProfile.contractProfiles
-      inheritanceGraph = systemProfile.inheritanceGraph
-
-      // console.log(msg)
-      assert(typeof contractProfiles === 'object')
-      assert(typeof inheritanceGraph === 'object')
-    })    
-    it('Includes a profile for each contract in the system', function() {
-      // convert the array of profiles to just an array of contract names
-      let profiledTokens = Object.keys(contractProfiles)
-
-      // check that each token in the system has been accounted for
-      for (let token of declaredTokens) {
-        let index = profiledTokens.indexOf(token)
-        assert(index !== -1, `${token} not found in system profile`)
-        // remove that token 
-        profiledTokens.splice(index, 1)
-      }
-      // TODO: this is failing, but rightfully so, because there are imports from beyond the directory.
-      assert(profiledTokens.length === 0, `System includes more profiles than expected: ${profiledTokens}`)
-    })
-  })
-
-  describe('derivedProfiler', function() {
-    it.skip('TBD test', function() {
-      
     })
   })
 })
